Move DbContext to the Data folder.
