+++ DesignPatternsInCSharp.Tests/Others/Repository/RepositoryTests.cs
@@ -1,4 +1,5 @@
 using DesignPatternsInCSharp.Others.Repository;
+using DesignPatternsInCSharp.Others.Repository.Data;
 using DesignPatternsInCSharp.Others.Repository.Interfaces;
 using DesignPatternsInCSharp.Others.Repository.Models;
 using Microsoft.EntityFrameworkCore;
