+++ DesignPatternsInCSharp/Others/Repository/Repository.cs
@@ -1,4 +1,5 @@
 using CommunityToolkit.Diagnostics;
+using DesignPatternsInCSharp.Others.Repository.Data;
 using DesignPatternsInCSharp.Others.Repository.Interfaces;
 using Microsoft.EntityFrameworkCore;
 using System.Linq.Expressions;
